Split the static Glimpse class into GlimpseConfiguration and GlimpseTimeline. Refactored other classes into appropriate files/namespaces.
