+++ source/Glimpse.Core/ClientScript/Metadata.cs
@@ -30,7 +30,7 @@ namespace Glimpse.Core.ClientScript
         /// Name of the <see cref="IResource" /> to link to.
         /// </returns>
         /// <remarks>
-        ///   <see cref="IResource" /> Uri generation is handled by implementations of <see cref="Glimpse.Core.Framework.ResourceEndpointConfiguration" />.
+        ///   <see cref="IResource" /> Uri generation is handled by implementations of <see cref="GlimpseTimeline.Core.Framework.ResourceEndpointConfiguration" />.
         /// </remarks>
         public string GetResourceName()
         {
