+++ source/Glimpse.Core/Extensibility/ITabLayout.cs
@@ -2,7 +2,7 @@ namespace Glimpse.Core.Extensibility
 {
     /// <summary>
     /// Definition for how the layout of a given object can be controlled. When
-    /// implementing a layout, <see cref="Glimpse.Core.Tab.Assist.TabLayout"/>
+    /// implementing a layout, <see cref="GlimpseTimeline.Core.Tab.Assist.TabLayout"/>
     /// can be used to make creating the required objects easier.
     /// </summary>
     public interface ITabLayout
