+++ source/Glimpse.Core/Glimpse.Core.csproj
@@ -118,7 +118,9 @@
     <Compile Include="Framework\Support\ExecutionBlockBase.cs" />
     <Compile Include="Framework\Support\IExecutionTask.cs" />
     <Compile Include="Framework\UriTemplateResourceEndpointConfiguration.cs" />
-    <Compile Include="Glimpse.cs" />
+    <Compile Include="GlimpseConfiguration.cs" />
+    <Compile Include="GlimpseTimeline.cs" />
+    <Compile Include="Message\TimelineMessage.cs" />
     <Compile Include="PreBodyTagInjectionStream.cs" />
     <Compile Include="Resource\ConfigurationScriptResource.cs">
       <SubType>Code</SubType>
@@ -190,7 +192,7 @@
     <Compile Include="Framework\GlimpseRequest.cs" />
     <Compile Include="Framework\GlimpseRequestHeaders.cs" />
     <Compile Include="Framework\IDiscoverableCollection.cs" />
-    <Compile Include="Framework\IGlimpseConfiguration.cs" />
+    <Compile Include="Framework\IConfiguration.cs" />
     <Compile Include="Framework\IGlimpseRuntime.cs" />
     <Compile Include="Constants.cs" />
     <Compile Include="Framework\IPersistenceStore.cs" />
@@ -213,7 +215,7 @@
     <Compile Include="Extensibility\ISerializer.cs" />
     <Compile Include="Extensibility\ITabContext.cs" />
     <Compile Include="Framework\IRequestResponseAdapter.cs" />
-    <Compile Include="Framework\GlimpseConfiguration.cs" />
+    <Compile Include="Framework\Configuration.cs" />
     <Compile Include="Framework\GlimpseRuntime.cs" />
     <Compile Include="Message\IMessage.cs" />
     <Compile Include="Message\ITimedMessage.cs" />
