+++ source/Glimpse.Core/Inspector/TraceInspector.cs
@@ -13,7 +13,7 @@ namespace Glimpse.Core.Inspector
         /// </summary>
         /// <param name="context">The context.</param>
         /// <remarks>
-        /// Executed during the <see cref="Glimpse.Core.Framework.IGlimpseRuntime.Initialize" /> phase of
+        /// Executed during the <see cref="GlimpseTimeline.Core.Framework.IGlimpseRuntime.Initialize" /> phase of
         /// system startup. Specifically, with the ASP.NET provider, this is wired to/implemented by the
         /// <c>System.Web.IHttpModule.Init</c> method.
         /// </remarks>
