+++ source/Glimpse.Test.Core/Glimpse.Test.Core.csproj
@@ -116,7 +116,7 @@
     <Compile Include="Framework\CastleDynamicProxyFactoryShould.cs" />
     <Compile Include="Framework\ApplicationPersistenceStoreShould.cs" />
     <Compile Include="Framework\MessageBrokerShould.cs" />
-    <Compile Include="Framework\GlimpseConfigurationShould.cs" />
+    <Compile Include="Framework\ConfigurationShould.cs" />
     <Compile Include="Framework\GlimpseMetadataShould.cs" />
     <Compile Include="Framework\GlimpseRuntimeShould.cs" />
     <Compile Include="Framework\InspectorContextShould.cs" />
